Created Folder structure for tests for app
