+++ src/app/tests/login_views_test.py
@@ -1,0 +1,73 @@
+# -*- coding: utf-8 -*-
+
+
+from django.test import TestCase
+from django.contrib.auth.models import User
+from django.conf import settings
+from django.urls import reverse
+
+
+from django.contrib.auth.models import User
+from django.conf import settings
+import os
+
+
+class LoginViewsTestCase(TestCase):
+
+    def initialise(self):
+        """ Create users"""
+        self.credentials = {'username':'testuser','password':'testpass' }
+        user = User.objects.create_user(**self.credentials)
+        user.is_staff = True
+        user.is_active = True
+        user.save()
+        self.credentials2 = {'username':'testuser2','password':'testpass2' }
+        user2 = User.objects.create_user(**self.credentials2)
+        user2.is_staff = False
+        user2.is_active = True
+        user2.save()
+        self.credentials3 = {'username':'testuser3','password':'testpass3' }
+        user3 = User.objects.create_user(**self.credentials3)
+        user3.is_staff = True
+        user3.is_active = False
+        user3.save()
+
+    def test_login(self):
+        """GET Request for login"""
+        resp = self.client.get(reverse("login"),follow=True,secure=True)
+        self.assertEqual(resp.status_code,200)
+        self.assertEqual(resp.redirect_chain,[])
+        self.assertIn("app/login.html",(i.name for i in resp.templates))
+        self.assertEqual(resp.resolver_match.func.__name__,"loginuser")
+
+    def test_postlogin(self):
+        """POST Request for index with different user types."""
+        self.initialise()
+        resp = self.client.post(reverse("login"),self.credentials,follow=True,secure=True)
+        self.assertEqual(resp.status_code,200)
+        self.assertNotEqual(resp.redirect_chain,[])
+        self.assertIn(settings.LOGIN_REDIRECT_URL, (i[0] for i in resp.redirect_chain))
+        self.assertTrue(resp.context['user'].is_active)
+        self.assertTrue(resp.context['user'].is_staff)
+        self.assertFalse(resp.context['user'].is_superuser)
+        self.client.get(reverse("logout"))
+
+        resp2 = self.client.post(reverse("login"),self.credentials2,follow=True,secure=True)
+        self.assertEqual(resp2.status_code,403)
+        self.assertEqual(resp2.redirect_chain,[])
+        self.assertFalse(resp2.context['user'].is_active)
+        self.assertFalse(resp2.context['user'].is_staff)
+        self.assertFalse(resp2.context['user'].is_superuser)
+        self.assertTrue('invalid' in resp2.context)
+        self.assertIn("app/login.html",(i.name for i in resp2.templates))
+        self.client.get(reverse("logout"))
+
+        resp3 = self.client.post(reverse("login"),self.credentials3,follow=True,secure=True)
+        self.assertEqual(resp3.status_code,403)
+        self.assertEqual(resp3.redirect_chain,[])
+        self.assertFalse(resp3.context['user'].is_active)
+        self.assertFalse(resp3.context['user'].is_staff)
+        self.assertFalse(resp3.context['user'].is_superuser)
+        self.assertTrue('invalid' in resp3.context)
+        self.assertIn("app/login.html",(i.name for i in resp3.templates))
+        self.client.get(reverse("logout"))
